waaletItemSericeTest / service/repository/serviceImpl

+++ src/main/java/com/wallet/walletapi/dto/WalletDto.java
@@ -1,5 +1,7 @@
 package com.wallet.walletapi.dto;
 
+import java.math.BigDecimal;
+
 import javax.validation.constraints.NotNull;
 
 import org.hibernate.validator.constraints.Length;
@@ -14,6 +16,6 @@ public class WalletDto {
 	@Length(min = 3, message = "O nome deve conter no minimo 3 carcter")
 	private String name;
 	@NotNull(message = "insira um valor pra a carteira")
-	private Long value;
+	private BigDecimal value;
 
 }
